pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = ""
authors = [
    {name = "gmcosta",email = "geu_costa@outlook.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi[standard] (>=0.121.2,<0.122.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "sqlalchemy[asyncio] (>=2.0.44,<3.0.0)",
    "alembic (>=1.17.1,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "pwdlib[argon2] (>=0.3.0,<0.4.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "psycopg[binary] (>=3.2.12,<4.0.0)"
]

[tool.poetry]
packages = [{include = "api", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.1,<10.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "taskipy (>=1.14.1,<2.0.0)",
    "ruff (>=0.14.5,<0.15.0)"
]

[tool.ruff]
line-length = 79

[tool.ruff.lint]
preview = true

[tool.ruff.format]
preview = true

[tool.pytest.ini_options]
pythonpath = "."

[tool.taskipy.tasks]
run = 'fastapi dev src/api/app.py'
test = 'pytest -s -x --cov=src/api -vv'